Applied relation between UserModel and ContactModel which have oneToMany and also applied authentication to contact APIs perticuler can only update, delete there own contact not others

+++ controllers/contactController.js
@@ -5,14 +5,14 @@ const Contact = require("../models/contactModel");
 // @route GET/api/contacts
 // @access public 
 const getContacts = asyncHandler(async(req , res) =>{
-    const contacts = await Contact.find();
+    const contacts = await Contact.find({user_id: req.user.id});
     res.status(200).json(contacts);
 });
 
 
 // @desc POST all contacts
 // @route POST/api/contacts
-// @access public 
+// @access private 
 const createContact = asyncHandler(async(req , res) =>{
     console.log("request body is : ",req.body);
     const {name , email , phone} = req.body;
@@ -25,6 +25,7 @@ const createContact = asyncHandler(async(req , res) =>{
         name,
         email,
         phone,
+        user_id:req.user.id,
     });
     res.status(201).json(contact);
 });
@@ -32,7 +33,7 @@ const createContact = asyncHandler(async(req , res) =>{
 
 // @desc Get all contacts
 // @route GET/api/contacts/:id
-// @access public 
+// @access private 
 const getContact = asyncHandler(async(req , res) =>{
     const contact = await Contact.findById(req.params.id);
     if(!contact){
@@ -45,12 +46,16 @@ const getContact = asyncHandler(async(req , res) =>{
 
 // @desc Update single contacts
 // @route PUT/api/contacts/:id
-// @access public 
+// @access private 
 const updateContact = asyncHandler(async(req , res) =>{
     const contact = await Contact.findById(req.params.id);
     if(!contact){
         res.status(404);
         throw new Error("Contact is not found");
+    }
+    if(contact.user_id.toString()!== req.user.id){
+        res.status(403);
+        throw new Error("user don't have permission to update other user contacts");
     }
     const updatedContact = await Contact.findByIdAndUpdate(
        req.params.id,
@@ -63,9 +68,16 @@ const updateContact = asyncHandler(async(req , res) =>{
 
 // @desc Delete single contacts
 // @route DELETE/api/contacts/:id
-// @access public 
+// @access private 
 const deleteContact = asyncHandler(async(req , res) =>{
     console.log("inside deleteContacts");
+    const contactcheck = await Contact.findById(req.params.id);
+    
+    if(contactcheck.user_id.toString()!== req.user.id){
+        res.status(403);
+        throw new Error("user don't have permission to update other user contacts");
+    }
+
     const contact = await Contact.findByIdAndDelete(req.params.id);
     console.log(contact);
     if(contact === null){
